requirements.txt
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.5
django-admin-list-filter-dropdown==1.0.3
django-celery-beat==2.5.0
django-json-widget==1.1.1
django-timezone-field==6.0
exceptiongroup==1.1.3
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.2
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.15
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
telegram-django-bot==1.1.0
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
